requirements.txt
# Equity in Focus - Capstone Project Requirements
# CIS 301 - Clark Atlanta CIS301
# Python 3.9+

# Backend (FastAPI)
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0

# Frontend (Streamlit)
streamlit==1.29.0
plotly==5.18.0
folium==0.15.0
streamlit-folium==0.15.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Cloud Storage - Google Cloud Storage integration
google-cloud-storage==2.10.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
requests==2.31.0

# Development
black==23.12.0
flake8==6.1.0

